Simple refactoring
