+++ case-study.md
@@ -1,0 +1,12 @@
+
+## Вникаем в детали системы, чтобы найти главные точки роста
+Для того, чтобы найти "точки роста" для оптимизации я воспользовался *инструментами, которыми вы воспользовались*
+
+Вот какие проблемы удалось найти и решить
+
+### Подготовка
+- простой рефакторинг кода, чтобы отделить тесты\классы и саму программу
+- возможность передавать название файла в программу
+
+### Ваша находка №1
+Проверить время
