+++ models/user.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+class User
+  attr_reader :attributes, :sessions
+
+  def initialize(attributes:, sessions:)
+    @attributes = attributes
+    @sessions = sessions
+  end
+end
